+++ chapter-1/seconds-old.rb
@@ -1,0 +1,19 @@
+puts ""
+puts "# = = = = = = = = = = "
+puts "# Chapter 1"
+puts "# File: seconds-old.rb"
+puts "# = = = = = = = = = = "
+puts ""
+
+puts "Constants:"
+puts "Age: 30"
+puts "Days within year: 365"
+puts "Hours within day: 24"
+puts "Months within year: 12"
+puts "Minutes within hour: 60"
+puts "Seconds within minute: 60"
+puts ""
+
+puts "How many seconds old:"
+puts 30*365*24*60*60
+puts ""