+++ CHANGELOG.md
@@ -10,3 +10,4 @@
 - chapter-1.rb [examples from chapter-1]
 - hours-within-year.rb [todo from chapter-1]
 - minutes-within-decade.rb [todo from chapter-1]
+- seconds-old.rb [todo from chapter-1]